Update command chain, so an agent sent back to his commander a tuple of 3 elements: a list of actions, a callback, and a list of parameters for the callback.
This mean that now, each agent need to work with a general. Then, the general will unpack the tuple, and process values inside.
If the first action of the list of action is valid, return it to pysc2.
Else, it will call the callback provided by the agent, and return NO_OP to pysc2. This way, agent do not need anymore to take care about validity of their actions.